Enter password when confirm transaction

+++ app/controllers/api/v1/pending_transactions_controller.rb
@@ -16,19 +16,23 @@ class Api::V1::PendingTransactionsController < ApplicationController
   def confirm
     pending = PendingTransaction.find_by(id: params[:id], user: current_user)
     if pending && current_user.authenticate_otp(params[:otp_code])
-      @service = KcoinService.new current_user
-      res = @service.create_transaction(pending.receiver, pending.amount)
-      if res
-        if res.code == "200"
-          pending.destroy
-          KcoinTransaction.syncing_transaction
-          render json:{ message: "Transaction created" }, status: 201
+      if current_user.valid_password? params[:password]
+        @service = KcoinService.new current_user
+        res = @service.create_transaction(pending.receiver, pending.amount)
+        if res
+          if res.code == "200"
+            pending.destroy
+            KcoinTransaction.syncing_transaction
+            render json:{ message: "Transaction created" }, status: 201
+          else
+            res = JSON.parse(res.body)
+            render json: { errors: res["message"] }, status: res.code.to_i
+          end
         else
-          res = JSON.parse(res.body)
-          render json: { errors: res["message"] }, status: res.code.to_i
+          render json: { errors: "Not enough money" }, status: 400
         end
       else
-        render json: { errors: "Not enough money" }, status: 400
+        render json: { errors: "Wrong password" }, status: 400
       end
     else
       render json: { errors: "Wrong OTP code" }, status: 400
